refactor: move onboarding tour interfaces to a separate file for improved organization and maintainability
